rework ADT and fix other things found not working (#25)
This PR re-implemented the ADT macro in a flavor closer to the rust enum. And we add quite a few new convenient functions for other purpose usages in meta-programming

# The new algebra data type

I realize it is not idiomatic to have variants to be an actual Julia type, which can be confusing. Instead, we accept the concept of variants, define a set of separate interfaces for them, and make the defined ADT act more like an enum type (but more general) in Julia. This seems more explicit and clear to work with the pattern match syntax.

The memory layout of the ADT is also re-designed, we move bits-type all together and use a 32-bit `primitive type` to denote the variant type. Then we let all the non-bits-type properties be of type `Any`. An extra type annotation is used inside `getproperty` to make it type-stable.

We still need to rewrite `struct_decons` to make pattern matching on the `call` syntax variant work. And further testing of the variant syntax is also required.

# The `guess_xxx` family

We put a set of new functions for guessing `type`, and `module` directly from the expression. This is quite useful when one wants to compare or parse expressions.

# Re-organized source files

The source files are reorganized as I find these files are growing large and should be separated now.

# Much better `@test_expr`

Now by rewriting `compare_expr`, we have a much better `@test_expr` that can correctly
return the line number of failed tests, and can locate the place inside the expression
causing the failure.