Support unparenthesised, single parameter, anonymous functions (#71)
These changes add support for anonymous functions of these shapes, all
of which are valid Julia, but to be parsed as `JLFunction`s:

- `x::Int->2x`
- `::Int -> 0`
- `x->2x`

`:(::)`-expressions with two arguments should only be split further when
the first argument is itself an `Expr`. That is generated by function
calls of the shape `(x,)::RT -> $body`, in which case the `(x,)` is the
`call` which should be split.

But Julia also generates two argument `:(::)`-expressions for functions of shape
`x::T -> $body`. In this case the first argument is `:x`, and the second
is `:T`. `:x` can't be split further, and so the whole matching `ex`,
`x::T`, should be returned as the only argument.

For functions of shape `::T -> $body`, an `:(::)`-expression with only
one argument, `:T` is generated, which should be handled in the same way
as above.

The last missing piece is the more common shape `x->$body`. These
functions only has a single symbol `:x` as the first argument of the
resulting `->`-expression. To handle this case, a new method for
`split_function_head` was added.